--- src/test/java/org/gridsuite/study/server/service/client/dynamicsimulation/DynamicSimulationClientTest.java
@@ -1,517 +1,0 @@
-/*
- * Copyright (c) 2022, RTE (http://www.rte-france.com)
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-package org.gridsuite.study.server.service.client.dynamicsimulation;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.github.tomakehurst.wiremock.client.WireMock;
-import org.gridsuite.study.server.RemoteServicesProperties;
-import org.gridsuite.study.server.dto.dynamicsimulation.DynamicSimulationStatus;
-import org.gridsuite.study.server.dto.dynamicsimulation.event.EventInfos;
-import org.gridsuite.study.server.service.StudyService;
-import org.gridsuite.study.server.service.client.AbstractWireMockRestClientTest;
-import org.gridsuite.study.server.service.client.timeseries.TimeSeriesClient;
-import org.gridsuite.study.server.service.dynamicsimulation.DynamicSimulationRestService;
-import org.gridsuite.study.server.utils.assertions.Assertions;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.MediaType;
-import org.springframework.web.client.HttpClientErrorException;
-import org.springframework.web.client.HttpServerErrorException;
-import org.springframework.web.client.RestTemplate;
-
-import java.util.List;
-import java.util.UUID;
-
-import static com.github.tomakehurst.wiremock.client.WireMock.equalTo;
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.gridsuite.study.server.StudyConstants.*;
-import static org.gridsuite.study.server.notification.NotificationService.HEADER_USER_ID;
-import static org.gridsuite.study.server.service.client.RestClient.DELIMITER;
-import static org.gridsuite.study.server.service.client.util.UrlUtil.buildEndPointUrl;
-import static org.gridsuite.study.server.service.dynamicsimulation.DynamicSimulationRestService.*;
-import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-
-class DynamicSimulationClientTest extends AbstractWireMockRestClientTest {
-
-    private static final UUID ROOT_NETWORK_UUID = UUID.randomUUID();
-    private static final UUID NETWORK_UUID = UUID.randomUUID();
-    private static final UUID REPORT_UUID = UUID.randomUUID();
-    private static final UUID NODE_UUID = UUID.randomUUID();
-    private static final UUID PARAMETERS_UUID = UUID.randomUUID();
-
-    private static final String DYNAMIC_SIMUALTION_RUN_BASE_URL = buildEndPointUrl("", API_VERSION, DYNAMIC_SIMULATION_END_POINT_RUN);
-    private static final String DYNAMIC_SIMULATION_RESULT_BASE_URL = buildEndPointUrl("", API_VERSION, DYNAMIC_SIMULATION_END_POINT_RESULT);
-    private static final String PARAMETERS_BASE_URL = buildEndPointUrl("", DYNAMIC_SIMULATION_API_VERSION, DYNAMIC_SIMULATION_END_POINT_PARAMETER);
-    private static final String PARAMETERS_JSON = "parametersJson";
-
-    private static final String VARIANT_ID = "variantId";
-
-    private static final UUID TIME_SERIES_UUID = UUID.randomUUID();
-    private static final UUID TIMELINE_UUID = UUID.randomUUID();
-    private static final UUID RESULT_UUID = UUID.randomUUID();
-    private static final UUID RESULT_NOT_FOUND_UUID = UUID.randomUUID();
-
-    private DynamicSimulationRestService dynamicSimulationRestService;
-
-    @Autowired
-    private RestTemplate restTemplate;
-
-    @Autowired
-    private ObjectMapper objectMapper;
-
-    @Autowired
-    private RemoteServicesProperties remoteServicesProperties;
-
-    @Autowired
-    private TimeSeriesClient timeSeriesClient;
-
-    @BeforeEach
-    void setup() {
-        // config client
-        remoteServicesProperties.setServiceUri("dynamic-simulation-server", initMockWebServer());
-        dynamicSimulationRestService = new DynamicSimulationRestService(remoteServicesProperties, restTemplate, objectMapper, timeSeriesClient);
-    }
-
-    @Test
-    void testGetProvider() {
-        String url = PARAMETERS_BASE_URL + DELIMITER + PARAMETERS_UUID + "/provider";
-
-        // --- Success --- //
-        String expectedProvider = DYNAWO_PROVIDER;
-
-        // configure mock server response
-        wireMockServer.stubFor(WireMock.get(WireMock.urlEqualTo(url))
-                .willReturn(WireMock.ok()
-                        .withBody(expectedProvider)
-                        .withHeader(HttpHeaders.CONTENT_TYPE, MediaType.TEXT_PLAIN_VALUE)
-                ));
-        // call service to test
-        String provider = dynamicSimulationRestService.getProvider(PARAMETERS_UUID);
-
-        // check result
-        assertThat(provider).isEqualTo(expectedProvider);
-
-        // --- Not Found --- //
-        // configure mock server response
-        wireMockServer.stubFor(WireMock.get(WireMock.urlEqualTo(url))
-                .willReturn(WireMock.notFound()));
-
-        // check result
-        assertThrows(
-            HttpClientErrorException.NotFound.class,
-            () -> dynamicSimulationRestService.getProvider(PARAMETERS_UUID)
-        );
-
-        // --- Error --- //
-        wireMockServer.stubFor(WireMock.get(WireMock.urlEqualTo(url))
-                .willReturn(WireMock.serverError()));
-
-        // check result
-        assertThrows(
-            HttpServerErrorException.class,
-            () -> dynamicSimulationRestService.getProvider(PARAMETERS_UUID)
-        );
-    }
-
-    @Test
-    void testGetParameters() {
-        String parametersJson = PARAMETERS_JSON;
-
-        String url = PARAMETERS_BASE_URL + DELIMITER + PARAMETERS_UUID;
-
-        // --- Success --- //
-        // configure mock server response
-        wireMockServer.stubFor(WireMock.get(WireMock.urlEqualTo(url))
-                .willReturn(WireMock.ok()
-                        .withBody(parametersJson)
-                        .withHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
-                ));
-        // call service to test
-        String resultParametersJson = dynamicSimulationRestService.getParameters(PARAMETERS_UUID);
-
-        // check result
-        assertThat(resultParametersJson).isEqualTo(parametersJson);
-
-        // --- Not Found --- //
-        // configure mock server response
-        wireMockServer.stubFor(WireMock.get(WireMock.urlEqualTo(url))
-                .willReturn(WireMock.notFound()));
-
-        // check result
-        assertThrows(
-            HttpClientErrorException.NotFound.class,
-            () -> dynamicSimulationRestService.getParameters(PARAMETERS_UUID)
-        );
-
-        // --- Error --- //
-        wireMockServer.stubFor(WireMock.get(WireMock.urlEqualTo(url))
-                .willReturn(WireMock.serverError()));
-
-        // check result
-        assertThrows(
-            HttpServerErrorException.class,
-            () -> dynamicSimulationRestService.getParameters(PARAMETERS_UUID)
-        );
-    }
-
-    @Test
-    void testCreateParameters() throws Exception {
-        String parameterJson = PARAMETERS_JSON;
-
-        // --- Success --- //
-        // configure mock server response
-        wireMockServer.stubFor(WireMock.post(WireMock.urlEqualTo(PARAMETERS_BASE_URL))
-                .withRequestBody(equalTo(parameterJson))
-                .willReturn(WireMock.ok()
-                        .withBody(objectMapper.writeValueAsString(PARAMETERS_UUID))
-                        .withHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
-                ));
-        // call service to test
-        UUID resultParametersUuid = dynamicSimulationRestService.createParameters(parameterJson);
-
-        // check result
-        assertThat(resultParametersUuid).isEqualTo(PARAMETERS_UUID);
-
-        // --- Error --- //
-        wireMockServer.stubFor(WireMock.post(WireMock.urlEqualTo(PARAMETERS_BASE_URL))
-                .withRequestBody(equalTo(parameterJson))
-                .willReturn(WireMock.serverError()));
-
-        // check result
-        assertThrows(
-            HttpServerErrorException.class,
-            () -> dynamicSimulationRestService.createParameters(parameterJson)
-        );
-    }
-
-    @Test
-    void testUpdateParameters() {
-        String parameterJson = PARAMETERS_JSON;
-
-        String url = PARAMETERS_BASE_URL + DELIMITER + PARAMETERS_UUID;
-
-        // --- Success --- //
-        // configure mock server response
-        wireMockServer.stubFor(WireMock.put(WireMock.urlEqualTo(url))
-                .withRequestBody(equalTo(parameterJson))
-                .willReturn(WireMock.ok()));
-        // call service to test
-        Assertions.assertThatNoException().isThrownBy(() -> dynamicSimulationRestService.updateParameters(PARAMETERS_UUID, parameterJson));
-
-        // --- Not Found --- //
-        // configure mock server response
-        wireMockServer.stubFor(WireMock.put(WireMock.urlEqualTo(url))
-                .withRequestBody(equalTo(parameterJson))
-                .willReturn(WireMock.notFound()));
-
-        // check result
-        assertThrows(
-            HttpClientErrorException.NotFound.class,
-            () -> dynamicSimulationRestService.updateParameters(PARAMETERS_UUID, parameterJson)
-        );
-
-        // --- Error --- //
-        wireMockServer.stubFor(WireMock.put(WireMock.urlEqualTo(url))
-                .withRequestBody(equalTo(parameterJson))
-                .willReturn(WireMock.serverError()));
-
-        // check result
-        assertThrows(
-            HttpServerErrorException.class,
-            () -> dynamicSimulationRestService.updateParameters(PARAMETERS_UUID, parameterJson)
-        );
-    }
-
-    @Test
-    void testDuplicateParameters() throws Exception {
-        UUID newParameterUuid = UUID.randomUUID();
-
-        // --- Success --- //
-        // configure mock server response
-        wireMockServer.stubFor(WireMock.post(WireMock.urlPathTemplate(PARAMETERS_BASE_URL + "/" + PARAMETERS_UUID + "/duplicate"))
-                    .willReturn(WireMock.ok()
-                        .withBody(objectMapper.writeValueAsString(newParameterUuid))
-                        .withHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
-                    ));
-        // call service to test
-        UUID resultParametersUuid = dynamicSimulationRestService.duplicateParameters(PARAMETERS_UUID);
-
-        // check result
-        assertThat(resultParametersUuid).isEqualTo(newParameterUuid);
-
-        // --- Not Found --- //
-        // configure mock server response
-        wireMockServer.stubFor(WireMock.post(WireMock.urlPathTemplate(PARAMETERS_BASE_URL + "/" + PARAMETERS_UUID + "/duplicate"))
-                .willReturn(WireMock.notFound()));
-
-        // check result
-        assertThrows(
-            HttpClientErrorException.NotFound.class,
-            () -> dynamicSimulationRestService.duplicateParameters(PARAMETERS_UUID)
-        );
-
-        // --- Error --- //
-        wireMockServer.stubFor(WireMock.post(WireMock.urlPathTemplate(PARAMETERS_BASE_URL + "/" + PARAMETERS_UUID + "/duplicate"))
-                .willReturn(WireMock.serverError()));
-
-        // check result
-        assertThrows(
-            HttpServerErrorException.class,
-            () -> dynamicSimulationRestService.duplicateParameters(PARAMETERS_UUID)
-        );
-    }
-
-    @Test
-    void testDeleteParameters() {
-        // configure mock server response
-        String url = PARAMETERS_BASE_URL + DELIMITER + PARAMETERS_UUID;
-        wireMockServer.stubFor(WireMock.delete(WireMock.urlEqualTo(url))
-                .willReturn(WireMock.ok()));
-        // call service to test
-        Assertions.assertThatNoException().isThrownBy(() -> dynamicSimulationRestService.deleteParameters(PARAMETERS_UUID));
-    }
-
-    @Test
-    void testCreateDefaultParameters() throws Exception {
-        String url = PARAMETERS_BASE_URL + "/default";
-
-        // --- Success --- //
-        // configure mock server response
-        wireMockServer.stubFor(WireMock.post(WireMock.urlEqualTo(url))
-                .willReturn(WireMock.ok()
-                        .withBody(objectMapper.writeValueAsString(PARAMETERS_UUID))
-                        .withHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
-                ));
-        // call service to test
-        UUID resultParametersUuid = dynamicSimulationRestService.createDefaultParameters();
-
-        // check result
-        assertThat(resultParametersUuid).isEqualTo(PARAMETERS_UUID);
-
-        // --- Error --- //
-        wireMockServer.stubFor(WireMock.post(WireMock.urlEqualTo(url))
-                .willReturn(WireMock.serverError()));
-
-        // check result
-        assertThrows(
-            HttpServerErrorException.class,
-            () -> dynamicSimulationRestService.createDefaultParameters()
-        );
-    }
-
-    @Test
-    void testRun() throws Exception {
-
-        // prepare parameters
-        List<EventInfos> events = ParameterTestUtils.getEventInfosList();
-
-        // configure mock server response for test case run - networks/{networkUuid}/run?
-        String url = DYNAMIC_SIMUALTION_RUN_BASE_URL + DELIMITER + NETWORK_UUID + DELIMITER + "run";
-        wireMockServer.stubFor(WireMock.post(WireMock.urlPathTemplate(url))
-                .withQueryParam(QUERY_PARAM_VARIANT_ID, equalTo(VARIANT_ID))
-                .withQueryParam(QUERY_PARAM_REPORT_UUID, equalTo(REPORT_UUID.toString()))
-                .withQueryParam(QUERY_PARAM_REPORTER_ID, equalTo(NODE_UUID.toString()))
-                .withQueryParam(QUERY_PARAM_REPORT_TYPE, equalTo(StudyService.ReportType.DYNAMIC_SIMULATION.reportKey))
-                .withQueryParam("parametersUuid", equalTo(PARAMETERS_UUID.toString()))
-                .withHeader(HEADER_USER_ID, equalTo("userId"))
-                .willReturn(WireMock.ok()
-                        .withBody(objectMapper.writeValueAsString(RESULT_UUID))
-                        .withHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
-                ));
-        // call service to test
-        UUID resultUuid = dynamicSimulationRestService.runDynamicSimulation(NODE_UUID, ROOT_NETWORK_UUID, NETWORK_UUID, VARIANT_ID,
-                REPORT_UUID, PARAMETERS_UUID, events, "userId", false);
-
-        // check result
-        assertThat(resultUuid).isEqualTo(RESULT_UUID);
-
-        // --- Error --- //
-        wireMockServer.stubFor(WireMock.post(WireMock.urlPathTemplate(url))
-                .withQueryParam(QUERY_PARAM_VARIANT_ID, equalTo("variantIdFailed"))
-                .withQueryParam(QUERY_PARAM_RECEIVER, equalTo("receiver"))
-                .withQueryParam(QUERY_PARAM_REPORT_UUID, equalTo(REPORT_UUID.toString()))
-                .withQueryParam(QUERY_PARAM_REPORTER_ID, equalTo(NODE_UUID.toString()))
-                .withQueryParam(QUERY_PARAM_REPORT_TYPE, equalTo(StudyService.ReportType.DYNAMIC_SIMULATION.reportKey))
-                .withHeader(QUERY_PARAM_DEBUG, equalTo("true"))
-                .withQueryParam("parametersUuid", equalTo(PARAMETERS_UUID.toString()))
-                .withHeader(HEADER_USER_ID, equalTo("userId"))
-                .willReturn(WireMock.serverError()));
-
-        // check result
-        assertThrows(
-            HttpClientErrorException.NotFound.class,
-            () -> dynamicSimulationRestService.runDynamicSimulation(NODE_UUID, ROOT_NETWORK_UUID, NETWORK_UUID, "variantIdFailed",
-                        REPORT_UUID, PARAMETERS_UUID, events, "userId", true)
-        );
-    }
-
-    @Test
-    void testGetTimeSeriesResult() throws Exception {
-
-        // configure mock server response for test get timeseries result uuid - results/{resultUuid}/timeseries
-        wireMockServer.stubFor(WireMock.get(WireMock.urlEqualTo(DYNAMIC_SIMULATION_RESULT_BASE_URL + DELIMITER + RESULT_UUID + DELIMITER + "timeseries"))
-                .willReturn(WireMock.ok()
-                        .withBody(objectMapper.writeValueAsString(TIME_SERIES_UUID))
-                        .withHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
-                        ));
-
-        UUID timeSeriesUuid = dynamicSimulationRestService.getTimeSeriesResult(RESULT_UUID);
-
-        // check result
-        assertThat(timeSeriesUuid).isEqualTo(TIME_SERIES_UUID);
-    }
-
-    @Test
-    void testGetTimeSeriesResultGivenBadUuid() {
-        // configure mock server response
-        wireMockServer.stubFor(WireMock.get(WireMock.urlEqualTo(DYNAMIC_SIMULATION_RESULT_BASE_URL + DELIMITER + RESULT_NOT_FOUND_UUID + DELIMITER + "timeseries"))
-                .willReturn(WireMock.notFound()
-                ));
-        assertThrows(
-            HttpClientErrorException.NotFound.class,
-            () -> dynamicSimulationRestService.getTimeSeriesResult(RESULT_NOT_FOUND_UUID)
-        );
-    }
-
-    @Test
-    void testGetTimelineResult() throws Exception {
-        // configure mock server response for test get timeline result - uuid results/{resultUuid}/timeline
-        wireMockServer.stubFor(WireMock.get(WireMock.urlEqualTo(DYNAMIC_SIMULATION_RESULT_BASE_URL + DELIMITER + RESULT_UUID + DELIMITER + "timeline"))
-                .willReturn(WireMock.ok()
-                        .withBody(objectMapper.writeValueAsString(TIMELINE_UUID))
-                        .withHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
-                ));
-        UUID timelineUuid = dynamicSimulationRestService.getTimelineRestResult(RESULT_UUID);
-        // check result
-        assertThat(timelineUuid).isEqualTo(TIMELINE_UUID);
-    }
-
-    @Test
-    void testGetTimelineResultGivenBadUuid() {
-        // configure mock server response
-        wireMockServer.stubFor(WireMock.get(WireMock.urlEqualTo(DYNAMIC_SIMULATION_RESULT_BASE_URL + DELIMITER + RESULT_NOT_FOUND_UUID + DELIMITER + "timeline"))
-                .willReturn(WireMock.notFound()
-                ));
-        assertThrows(
-            HttpClientErrorException.NotFound.class,
-            () -> dynamicSimulationRestService.getTimelineResult(RESULT_NOT_FOUND_UUID)
-        );
-    }
-
-    @Test
-    void testGetStatus() throws Exception {
-
-        // configure mock server response for test get status result - results/{resultUuid}/status
-        wireMockServer.stubFor(WireMock.get(WireMock.urlEqualTo(DYNAMIC_SIMULATION_RESULT_BASE_URL + DELIMITER + RESULT_UUID + DELIMITER + "status"))
-                .willReturn(WireMock.ok()
-                        .withBody(objectMapper.writeValueAsString(DynamicSimulationStatus.CONVERGED))
-                        .withHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
-                ));
-        DynamicSimulationStatus status = dynamicSimulationRestService.getStatus(RESULT_UUID);
-        // check result
-        assertThat(status).isEqualTo(DynamicSimulationStatus.CONVERGED);
-    }
-
-    @Test
-    void testGetStatusGivenBadUuid() {
-        // configure mock server response
-        wireMockServer.stubFor(WireMock.get(WireMock.urlEqualTo(DYNAMIC_SIMULATION_RESULT_BASE_URL + DELIMITER + RESULT_NOT_FOUND_UUID + DELIMITER + "status"))
-                .willReturn(WireMock.notFound()
-                ));
-        assertThrows(
-            HttpClientErrorException.NotFound.class,
-            () -> dynamicSimulationRestService.getStatus(RESULT_NOT_FOUND_UUID)
-        );
-    }
-
-    @Test
-    void testInvalidateStatus() throws Exception {
-
-        // configure mock server response for test get status result - results/{resultUuid}/invalidate-status
-        wireMockServer.stubFor(WireMock.put(WireMock.urlMatching(DYNAMIC_SIMULATION_RESULT_BASE_URL + DELIMITER + "invalidate-status" + ".*"))
-                .withQueryParam("resultUuid", equalTo(RESULT_UUID.toString()))
-                .willReturn(WireMock.ok()
-                        .withBody(objectMapper.writeValueAsString(List.of(RESULT_UUID)))
-                        .withHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
-                ));
-        // test service
-        assertDoesNotThrow(() -> dynamicSimulationRestService.invalidateStatus(List.of(RESULT_UUID)));
-    }
-
-    @Test
-    void testInvalidateStatusGivenBadUuid() {
-        // configure mock server response for test get status result - results/{resultUuid}/invalidate-status
-        wireMockServer.stubFor(WireMock.get(WireMock.urlMatching(DYNAMIC_SIMULATION_RESULT_BASE_URL + DELIMITER + "invalidate-status" + ".*"))
-                .withQueryParam("resultUuid", equalTo(RESULT_NOT_FOUND_UUID.toString()))
-                .willReturn(WireMock.notFound()
-                ));
-        assertThrows(
-            HttpClientErrorException.NotFound.class,
-            () -> dynamicSimulationRestService.invalidateStatus(List.of(RESULT_NOT_FOUND_UUID))
-        );
-    }
-
-    @Test
-    void testDeleteResult() {
-        // configure mock server response for test delete result.
-        wireMockServer.stubFor(WireMock.delete(WireMock.urlEqualTo(DYNAMIC_SIMULATION_RESULT_BASE_URL + "?resultsUuids=" + RESULT_UUID))
-                .willReturn(WireMock.ok()
-                        .withHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
-                ));
-        // test service
-        assertDoesNotThrow(() -> dynamicSimulationRestService.deleteResults(List.of(RESULT_UUID)));
-    }
-
-    @Test
-    void testDeleteResults() throws Exception {
-        // configure mock server response for test delete all results - results/
-        wireMockServer.stubFor(WireMock.delete(WireMock.urlEqualTo(DYNAMIC_SIMULATION_RESULT_BASE_URL + "?resultsUuids"))
-            .willReturn(WireMock.ok()
-                .withHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
-            ));
-        dynamicSimulationRestService.deleteAllResults();
-
-        // configure mock server response for test result count - supervision/results-count
-        String resultCountEndPointUrl = buildEndPointUrl("", API_VERSION, DYNAMIC_SIMULATION_END_POINT_RESULT_COUNT);
-        wireMockServer.stubFor(WireMock.get(WireMock.urlEqualTo(resultCountEndPointUrl))
-            .willReturn(WireMock.ok()
-                .withBody(objectMapper.writeValueAsString(0))
-                .withHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
-            ));
-        Integer resultCount = dynamicSimulationRestService.getResultsCount();
-
-        // check result
-        assertThat(resultCount).isZero();
-    }
-
-    @Test
-    void testGetProviders() {
-        String url = buildEndPointUrl("", DYNAMIC_SIMULATION_API_VERSION, "providers");
-        String providers = "[\"Dynawo\"]";
-        wireMockServer.stubFor(WireMock.get(WireMock.urlEqualTo(url))
-            .willReturn(WireMock.ok().withBody(providers).withHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)));
-
-        assertThat(dynamicSimulationClient.getProviders()).isEqualTo(providers);
-        wireMockServer.verify(WireMock.getRequestedFor(WireMock.urlEqualTo(url)));
-    }
-
-    @Test
-    void testDownloadDebugFile() throws Exception {
-        String body = "{\"debug\":true}";
-        String url = DYNAMIC_SIMULATION_RESULT_BASE_URL + DELIMITER + RESULT_UUID + DELIMITER + "download-debug-file";
-        wireMockServer.stubFor(WireMock.get(WireMock.urlEqualTo(url)).willReturn(WireMock.ok().withBody(body)));
-
-        var response = dynamicSimulationClient.downloadDebugFile(RESULT_UUID);
-        assertThat(response.getStatusCode().is2xxSuccessful()).isTrue();
-        assertThat(new String(response.getBody().getInputStream().readAllBytes(), java.nio.charset.StandardCharsets.UTF_8)).isEqualTo(body);
-        wireMockServer.verify(WireMock.getRequestedFor(WireMock.urlEqualTo(url)));
-    }
-
-}
