+++ src/test/java/org/gridsuite/study/server/service/dynamicmargincalculation/DynamicMarginCalculationClientTest.java
@@ -431,7 +431,7 @@ class DynamicMarginCalculationClientTest extends AbstractWireMockRestClientTest 
         wireMockServer.stubFor(WireMock.get(WireMock.urlEqualTo(url))
             .willReturn(WireMock.ok().withBody(providers).withHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)));
 
-        assertThat(dynamicMarginCalculationClient.getProviders()).isEqualTo(providers);
+        assertThat(dynamicMarginCalculationRestService.getProviders()).isEqualTo(providers);
         wireMockServer.verify(WireMock.getRequestedFor(WireMock.urlEqualTo(url)));
     }
 
@@ -441,7 +441,7 @@ class DynamicMarginCalculationClientTest extends AbstractWireMockRestClientTest 
         String url = RESULT_BASE_URL + DELIMITER + RESULT_UUID + DELIMITER + "download-debug-file";
         wireMockServer.stubFor(WireMock.get(WireMock.urlEqualTo(url)).willReturn(WireMock.ok().withBody(body)));
 
-        var response = dynamicMarginCalculationClient.downloadDebugFile(RESULT_UUID);
+        var response = dynamicMarginCalculationRestService.downloadDebugFile(RESULT_UUID);
         assertThat(response.getStatusCode().is2xxSuccessful()).isTrue();
         assertThat(new String(response.getBody().getInputStream().readAllBytes(), java.nio.charset.StandardCharsets.UTF_8)).isEqualTo(body);
         wireMockServer.verify(WireMock.getRequestedFor(WireMock.urlEqualTo(url)));
